requirements.txt
torch==1.2.0
torchvision==0.4.0
gym==0.12.0
atari-py==0.1.4
scikit-image==0.15.0
matplotlib>=3.0.2
h5py>=3.1.0
numpy>=1.19.5
scipy>=0.19.1
Pillow>=8.2.0